test-requirements.txt
coverage>=4.0
flake8>=6.1.0  # MIT License (MIT)
mock>=3.0.5,<4.0.0  # BSD
pytest==7.0.1,<8.0.0  # MIT License (MIT)
pytest-timer==0.0.11  # MIT License (MIT)
pyyaml>=6.0.0,<7.0.0  # MIT
